# openapi.yaml
openapi: 3.0.3
info:
  title: Point of Sale (POS) System API
  version: 1.0.0
  description: >
    Comprehensive REST API documentation for POS System Backend featuring multi-branch management,
    real-time analytics, inventory control, cashier shift management, refund processing, and payment integration.
  contact:
    name: POS System Support
    email: support@pos-system.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local Development Server
security:
  - BearerAuth: []
paths:
...
  /api/branch/{id}:
    get:
      tags:
        - Branch Management
      summary: Get Branch by ID
      description: Retrieves branch details by branch ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved branch details
    put:
      tags:
        - Branch Management
      summary: Update Branch
      description: Updates an existing branch's details by branch ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BranchDTO'
      responses:
        '200':
          description: Branch successfully updated
    delete:
      tags:
        - Branch Management
      summary: Delete Branch
      description: Deletes a branch by branch ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Branch successfully deleted
...
  /api/refunds/shift/{shiftId}:
    get:
      tags:
        - Refund Management
      summary: Get Refunds by Shift Report ID
      description: Retrieves all refunds processed during a specific cashier shift report.
      parameters:
        - name: shiftId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved shift refunds
...
  /api/shift-reports/{shiftReportId}:
    get:
      tags:
        - Shift Report Management
      summary: Get Shift Report by ID
      description: Retrieves detailed shift report details by shift report ID.
      parameters:
        - name: shiftReportId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved shift report
...
  /api/shift-reports/cashier/{cashierId}:
    get:
      tags:
        - Shift Report Management
      summary: Get Shift Reports by Cashier ID
      description: Retrieves all past and current shift reports for a specific cashier.
      parameters:
        - name: cashierId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved cashier shift reports
  /api/shift-reports/branch/{branchId}:
    get:
      tags:
        - Shift Report Management
      summary: Get Shift Reports by Branch ID
      description: Retrieves all shift reports associated with a specific branch.
      parameters:
        - name: branchId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved branch shift reports
...
  /api/shift-reports/cashier/{cashierId}/date:
    get:
      tags:
        - Shift Report Management
      summary: Get Shift Report by Cashier and Date
      description: Retrieves a cashier's shift report for a specific date and time.
      parameters:
        - name: cashierId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Successfully retrieved matching shift report

  # ---------------------------------------------------------------------------
  # Branch Analytics
  # ---------------------------------------------------------------------------
...
  /api/store-analytics/{storeAdminId}/sales-trends:
    get:
      tags:
        - Store Analytics
      summary: Get Sales Trends
      description: Retrieves time-series sales trend data for a store admin filtered by period (daily, weekly, monthly, yearly).
      parameters:
        - name: storeAdminId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: period
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved sales trends
  /api/store-analytics/{storeAdminId}/overview:
    get:
      tags:
        - Store Analytics
      summary: Get Store Overview Metrics
      description: Retrieves overall store metrics (total revenue, total orders, active branches, total customers) for a store admin.
      parameters:
        - name: storeAdminId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved store overview
  /api/store-analytics/{storeAdminId}/monthly-sales:
    get:
      tags:
        - Store Analytics
      summary: Get Monthly Sales Graph
      description: Retrieves monthly sales data points for graph plotting.
      parameters:
        - name: storeAdminId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved monthly sales data points
  /api/store-analytics/{storeAdminId}/daily-sales:
    get:
      tags:
        - Store Analytics
      summary: Get Daily Sales Graph
      description: Retrieves daily sales data points for graph plotting.
      parameters:
        - name: storeAdminId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved daily sales data points
  /api/store-analytics/{storeAdminId}/sales-payment-method:
    get:
      tags:
        - Store Analytics
      summary: Get Sales by Payment Method
      description: Retrieves store sales breakdown by payment method.
      parameters:
        - name: storeAdminId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved payment insights
  /api/store-analytics/{storeAdminId}/branch-sales:
    get:
      tags:
        - Store Analytics
      summary: Get Sales by Branch
      description: Retrieves sales comparison data across all branches owned by a store admin.
      parameters:
        - name: storeAdminId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved branch sales comparison
  /api/store-analytics/{storeAdminId}/branch-payment-breakdown:
    get:
      tags:
        - Store Analytics
      summary: Get Branch Payment Breakdown
      description: Retrieves payment method breakdown across store branches.
      parameters:
        - name: storeAdminId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved branch payment breakdown
  /api/store-analytics/{storeAdminId}/branch-performance:
    get:
      tags:
        - Store Analytics
      summary: Get Branch Performance
      description: Retrieves top performing and underperforming branches for a store admin.
      parameters:
        - name: storeAdminId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved branch performance report
  /api/store-analytics/{storeAdminId}/alert:
    get:
      tags:
        - Store Analytics
      summary: Get Store Alerts
      description: Retrieves store-level operational alerts (low stock items, refund spikes, inactive cashiers, branches with no sales).
      parameters:
        - name: storeAdminId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved store alerts

  # ---------------------------------------------------------------------------
  # Super Admin Dashboard
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your JWT Bearer token obtained from `/auth/login` or `/auth/signup`
  schemas:
...
    BranchDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        storeId:
          type: integer
          format: int64
...